+++ src/classes/UI5Classes/UI5Parser/UIClass/StandardUIClass.ts
@@ -253,6 +253,8 @@ export class StandardUIClass extends AbstractUIClass {
 			i++;
 		}
 
+		textWithoutTags = textWithoutTags.trim();
+
 		return textWithoutTags;
 	}