+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.12.16 (21-03-2021)
 * Support for controller extensions in manifest added
 * Completion items for overriding methods/fields added
+* Data types to JS method parameters Hover added
 
 ## 0.12.15 (20-03-2021)
 * Update keywords